backend/requirements.txt
fastapi
uvicorn
google-generativeai
weaviate-client
python-multipart
python-dotenv
groq
